+++ OpenSim/Region/CoreModules/World/Terrain/PaintBrushes/RaiseSphere.cs
@@ -51,7 +51,7 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
 
                     // Calculate a cos-sphere and add it to the heighmap
                     double r = Math.Sqrt(dx2 + (y - ry) * (y - ry));
-                    double distancefactor = Math.Cos(r * size);
+                    float distancefactor = (float)Math.Cos(r * size);
                     if (distancefactor > 0.0)
                         map[x, y] += distancefactor * strength;
                 }
