+++ OpenSim/Region/ScriptEngine/Shared/Api/Implementation/OSSL_Api.cs
@@ -599,7 +599,7 @@ namespace OpenSim.Region.ScriptEngine.Shared.Api
 
             if (World.Permissions.CanTerraformLand(m_host.OwnerID, new Vector3(x, y, 0)))
             {
-                World.Heightmap[x, y] = val;
+                World.Heightmap[x, y] = (float)val;
                 return 1;
             }
             else
